clean up commit ad101971f171c43cc8a28fbe44479cbe06be8db7 by making handling of zero values consistent
unit test results:

[INFO] ------------------------------------------------------------------------
[INFO] Reactor Summary:
[INFO]
[INFO] java root ......................................... SUCCESS [0.579s]
[INFO] jcommon util ...................................... SUCCESS [7.904s]
[INFO] jcommon collections foundation package ............ SUCCESS [1.730s]
[INFO] jcommon data package .............................. SUCCESS [7.420s]
[INFO] jcommon logging ................................... SUCCESS [3.571s]
[INFO] jcommon collections package ....................... SUCCESS [14.681s]
[INFO] jcommon config .................................... SUCCESS [8.287s]
[INFO] jcommon lifecycle ................................. SUCCESS [3.571s]
[INFO] shared testing utilities .......................... SUCCESS [0.559s]
[INFO] jcommon concurrency ............................... SUCCESS [13.408s]
[INFO] jcommon stats package ............................. SUCCESS [1:03.886s]
[INFO] zookeeper libraries ............................... SUCCESS [1.601s]
[INFO] jcommon memory package ............................ SUCCESS [3.577s]
[INFO] jcommon-all ....................................... SUCCESS [0.054s]
[INFO] ------------------------------------------------------------------------
[INFO] BUILD SUCCESS
[INFO] ------------------------------------------------------------------------
[INFO] Total time: 1:28.108s (Wall Clock)
[INFO] Finished at: Sat Mar 16 23:53:26 PDT 2013
[INFO] Final Memory: 47M/314M
[INFO] ------------------------------------------------------------------------

+++ data/src/main/java/com/facebook/data/types/StringDatum.java
@@ -46,11 +46,11 @@ public class StringDatum implements Datum {
     }
 
     try {
-      // TODO this treats -0, 00, 000, etc., as false; is this what we want?
-      // (and notably does not think 0.0 is false)
-      return asByte() != 0;
+      // treat all values that are effectively 0 as false. This includes +/- 0, +/-0.0, any
+      // sequence of +/- 0s ("0000"), and so on
+      return (int)(asDouble()) != 0;
     } catch (NumberFormatException e) {
-      // a value that isn't a number, but isn't "" or "false" is still 
+      // a value that isn't a number, but isn't "" or "false" is still
       // considered true, so ignore this
     }
